+++ app/Http/Controllers/UserController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
-use App\Models\Invite;
+// use App\Models\Invite;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\URL;
@@ -12,6 +12,10 @@ use Illuminate\Support\Facades\Hash;
 use App\Notifications\InviteNotification;
 use Illuminate\Support\Facades\Notification;
 use Symfony\Component\HttpFoundation\Response;
+use App\Models\Role;
+use App\Models\Authentication\Permission;
+use App\Models\Authentication\RolePermission;
+use App\Models\Authentication\UserRole;
 
 class UserController extends Controller
 {
@@ -23,6 +27,8 @@ class UserController extends Controller
     
     public function index()
     {
+        $this->authorize('viewAny', User::class);
+
         $users = User::all();
         return view('users.index', compact('users'));
     }
@@ -34,7 +40,11 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('users.create');
+        $this->authorize('create', User::class);
+
+        $roles = Role::get();
+        $permissions = Permission::get();
+        return view('users.create', compact('roles', 'permissions'));
     }
 
     /**
@@ -45,28 +55,39 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
 
         $request->validate([
             'first_name' => ['required', 'string', 'max:255'],
-            'middle_name' => ['string', 'max:255'],
+            'middle_name' => ['max:255'],
             'last_name' => ['required', 'string', 'max:255'],
-            'date_of_birth' => ['required'],
+            'birthdate' => ['required'],
             'email' => ['required', 'email', 'max:255', 'unique:users'],
             'password' => ['required'],
         ]);
 
-        User::create([
+        $user = User::create([
             'first_name' => $request->input('first_name'),
             'middle_name' => $request->input('middle_name'),
             'last_name' => $request->input('last_name'),
             'suffix' => $request->input('suffix') ?? null,
-            'date_of_birth' => $request->input('date_of_birth'),
-            'role_id' => $request->input('role_id'),
+            'date_of_birth' => $request->input('birthdate'),
+            'role_id' => 1,
             'email' => $request->input('email'),
             'password' => Hash::make($request->input('password')),
         ]);
 
-        return redirect()->route('users.index')->with('success','User added successfully.');
+        $user_id = $user->id;
+
+        $roles = $request->input('roles');
+        foreach($roles as $role){
+            $user->userrole()->create([
+                'user_id' => $user_id,
+                'role_id' => $role,
+            ]);
+        }
+
+        return redirect()->route('admin.users.create')->with('add_user_success','User added successfully.');
     }
 
     /**
@@ -77,6 +98,9 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize('view', User::class);
+
+        $users = User::all();
         return view('users.show', compact('user'));
     }
 
@@ -88,7 +112,16 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        return view('users.edit', compact('user'));
+        $this->authorize('update', User::class);
+
+        // $rolehaspermissions = RolePermission::join('user_roles', 'user_roles.role_id', '=', 'role_permissions.role_id')
+        //                                 ->where('user_roles.user_id', '=', $user->id)
+        //                                 ->select('role_permissions.role_id')
+        //                                 ->get();
+        $roles = Role::get();
+        $yourroles = $user->userrole()->pluck('role_id')->all();
+        $permissions = Permission::get();
+        return view('users.edit', compact('user', 'roles', 'permissions', 'yourroles'));
     }
 
     /**
@@ -100,22 +133,47 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
-        $request->validate([
-            'first_name' => ['required', 'string', 'max:255'],
-            'middle_name' => ['required', 'string', 'max:255'],
-            'last_name' => ['required', 'string', 'max:255'],
-            'date_of_birth' => ['required'],
-        ]);
-        $user->update([
-            'first_name' => $request->input('first_name'),
-            'middle_name' => $request->input('middle_name'),
-            'last_name' => $request->input('last_name'),
-            'suffix' => $request->input('suffix') ?? null,
-            'date_of_birth' => $request->input('date_of_birth'),
-            'role_id' => $request->input('role_id'),
-        ]);
+        $this->authorize('update', User::class);
 
-        return redirect()->route('users.index')->with('success','User updated successfully.');
+        $checkedroles = $request->input('roles');
+
+        $allroles = $user->userrole()->pluck('role_id')->all();
+        $trashedroles = $user->userrole()->onlyTrashed()->pluck('role_id')->all();
+
+        if ($checkedroles == null) {
+            $user->userrole()->delete();
+        }
+
+        else {
+            foreach ($checkedroles as $checkedrole){
+
+                if ($user->userrole($checkedrole)) {
+                    if (in_array($checkedrole, $trashedroles)) {
+                        $user->userrole()->where('role_id', $checkedrole)->restore();
+                    }
+                }
+
+                if (!(in_array($checkedrole, $allroles))) {
+                    if (in_array($checkedrole, $trashedroles)) {
+                        $user->userrole()->where('role_id', $checkedrole)->restore();
+                    }
+                    else {
+                        $user->userrole()->create([
+                            'user_id' => $user->id,
+                            'role_id' => $checkedrole
+                        ]);
+                    }
+                }
+            }
+
+            foreach ($allroles as $role) {
+                if (!(in_array($role, $checkedroles))) {
+                    $user->userrole()->where('role_id', $role)->delete();
+                }
+            }
+        }
+
+        return redirect()->route('admin.users.index')->with('edit_user_success','User record updated successfully.');
     }
 
     /**
@@ -126,9 +184,11 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', User::class);
+
         $user->delete();
 
-        return redirect()->route('users.index')->with('success', 'User deleted successfully');
+        return redirect()->route('admin.users.index')->with('edit_user_success', 'User record deleted successfully');
     }
 
     // public function invite(){
